// Frontend/src/app/api/student.ts
export class StudentModel implements IStudentModel {
    id?: number;
    username?: string | undefined;
    firstName?: string | undefined;
    lastName?: string | undefined;
    age?: number;
    career?: string | undefined;

    constructor(data?: IStudentModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.username = _data["username"];
            this.firstName = _data["firstName"];
            this.lastName = _data["lastName"];
            this.age = _data["age"];
            this.career = _data["career"];
        }
    }

    static fromJS(data: any): StudentModel {
        data = typeof data === 'object' ? data : {};
        let result = new StudentModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["username"] = this.username;
        data["firstName"] = this.firstName;
        data["lastName"] = this.lastName;
        data["age"] = this.age;
        data["career"] = this.career;
        return data; 
    }
}

export interface IStudentModel {
    id?: number;
    username?: string | undefined;
    firstName?: string | undefined;
    lastName?: string | undefined;
    age?: number;
    career?: string | undefined;
}